test: Rename 'overwrite' to 'no-overwriting'
