+++ tests/Fixtures/no-overwriting-library/composer.json
@@ -25,7 +25,7 @@
                     "files/mix/bin/hello-python"
                 ]
             },
-            "overwrite": {
+            "no-overwriting": {
                 "type": "file",
                 "url": "http://localhost:8000/file/ipsum",
                 "path": "library-file.txt"
