+++ tests/Fixtures/no-overwriting-library/library-file.txt
